Simplified the Travis build matrix
Testing the older releases and the dev versions of vendors separately
and against all PHP versions is wasting the Travis resources and making
builds longer given that the number of concurrent jobs is limited.
Testing these libraries on each PHP version is not needed (they already
do it) and they can be combined in the same build.
This only going down from 20 jobs per build to 6.