Remove useless suggestions
Suggestions are meant to link to packages which can enhance the features of a package.

None of these suggestions are about enhancing features of Pagerfanta. If the project does not already use one of these packages (which means the suggestions is already hidden by composer), there is no reason to use the corresponding adapter.